move compass home to compass
