--- packages/compass-home/.browserslistrc
@@ -1,1 +1,0 @@
-Electron >= 6
