--- packages/compass-home/.depcheckrc
@@ -1,8 +1,0 @@
-ignores:
-  - '@electron/remote'
-  - '@mongodb-js/prettier-config-compass'
-  - '@mongodb-js/tsconfig-compass'
-  - 'eventemitter3'
-  - 'resolve'
-ignore-patterns:
-  - 'dist'
