--- packages/compass-home/.eslintignore
@@ -1,2 +1,0 @@
-.nyc-output
-dist
