--- packages/compass-home/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['@mongodb-js/eslint-config-compass/plugin'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['./tsconfig-lint.json'],
-  },
-};
