--- packages/compass-home/.gitignore
@@ -1,1 +1,0 @@
-src/assets/fonts/akzid*
