--- packages/compass-home/.npmignore
@@ -1,19 +1,0 @@
-.babelrc
-.eslintrc
-.editorconfig
-.nycrc
-.travis.yml
-config/
-electron/
-dist/index.html
-karma.conf.js
-mocha-webpack.opts
-scripts/
-src/
-test/
-.nyc_output/
-coverage/
-.github/
-
-scripts/**/*
-!scripts/download-fonts.js
