--- packages/compass-home/.nycrc
@@ -1,6 +1,0 @@
-{
-  "exclude": ["config", ".tmp", "test"],
-  "reporter": ["text", "html"],
-  "sourceMap": false,
-  "instrument": false
-}
