--- packages/compass-home/.prettierignore
@@ -1,3 +1,0 @@
-dist
-.nyc_output
-coverage
