--- packages/compass-home/.prettierrc.json
@@ -1,1 +1,0 @@
-"@mongodb-js/prettier-config-compass"
