--- packages/compass-home/README.md
@@ -1,25 +1,0 @@
-# compass-home
-
-The Compass plugin responsible for gluing together other plugins:
-
-- [compass-aggregations][compass-aggregations]
-- [compass-collection][compass-collection]
-- [compass-crud][compass-crud]
-- [compass-database][compass-database]
-- [compass-databases-collections][compass-databases-collections]
-- [compass-deployment-awareness][compass-deployment-awareness]
-- [compass-export-to-language][compass-export-to-language]
-- [compass-field-store][compass-field-store]
-- [compass-find-in-page][compass-find-in-page]
-- [compass-import-export][compass-import-export]
-- [compass-indexes][compass-indexes]
-- [compass-instance][compass-instance]
-- [compass-query-bar][compass-query-bar]
-- [compass-query-history][compass-query-history]
-- [compass-schema-validation][compass-schema-validation]
-- [compass-server-version][compass-server-version]
-- [compass-serverstats][compass-serverstats]
-- [compass-shell][compass-shell]
-- [compass-sidebar][compass-sidebar]
-- [compass-ssh-tunnel-status][compass-ssh-tunnel-status]
-- [compass-status][compass-status]
