--- packages/compass-home/package.json
@@ -1,142 +1,0 @@
-{
-  "name": "@mongodb-js/compass-home",
-  "productName": "Home plugin",
-  "version": "7.0.1",
-  "apiVersion": "3.0.0",
-  "description": "Home",
-  "main": "dist/index.js",
-  "compass:main": "src/index.ts",
-  "types": "dist/index.d.ts",
-  "exports": {
-    "browser": "./dist/browser.js",
-    "require": "./dist/index.js"
-  },
-  "compass:exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "webpack": "webpack-compass",
-    "postcompile": "tsc --emitDeclarationOnly",
-    "compile": "npm run webpack -- --mode production",
-    "prettier": "prettier",
-    "test": "mocha",
-    "test-electron": "xvfb-maybe electron-mocha \"./src/**/*.spec.tsx\" --no-sandbox",
-    "check": "npm run lint && npm run depcheck",
-    "eslint": "eslint",
-    "prepublishOnly": "npm run compile && compass-scripts check-exports-exist",
-    "lint": "npm run eslint . && npm run prettier -- --check .",
-    "depcheck": "compass-scripts check-peer-deps && depcheck",
-    "test-cov": "nyc --compact=false --produce-source-map=false -x \"**/*.spec.*\" --reporter=lcov --reporter=text --reporter=html npm run test",
-    "test-ci": "npm run test-cov",
-    "test-ci-electron": "npm run test-electron",
-    "bootstrap": "npm run compile",
-    "reformat": "npm run eslint . -- --fix && npm run prettier -- --write ."
-  },
-  "license": "SSPL",
-  "dependencies": {
-    "@mongodb-js/atlas-service": "^0.15.1",
-    "@mongodb-js/compass-aggregations": "^9.26.1",
-    "@mongodb-js/compass-app-stores": "^7.10.1",
-    "@mongodb-js/compass-collection": "^4.23.1",
-    "@mongodb-js/compass-components": "^1.22.1",
-    "@mongodb-js/compass-connections": "^1.25.1",
-    "@mongodb-js/compass-crud": "^13.24.1",
-    "@mongodb-js/compass-databases-collections": "^1.23.1",
-    "@mongodb-js/compass-explain-plan": "^6.24.1",
-    "@mongodb-js/compass-export-to-language": "^9.0.1",
-    "@mongodb-js/compass-field-store": "^9.0.19",
-    "@mongodb-js/compass-find-in-page": "^4.22.1",
-    "@mongodb-js/compass-import-export": "^7.23.1",
-    "@mongodb-js/compass-indexes": "^5.23.1",
-    "@mongodb-js/compass-logging": "^1.2.14",
-    "@mongodb-js/compass-query-bar": "^8.25.1",
-    "@mongodb-js/compass-saved-aggregations-queries": "^1.24.1",
-    "@mongodb-js/compass-schema": "^6.25.1",
-    "@mongodb-js/compass-schema-validation": "^6.24.1",
-    "@mongodb-js/compass-serverstats": "^16.23.1",
-    "@mongodb-js/compass-settings": "^0.26.1",
-    "@mongodb-js/compass-shell": "^3.23.1",
-    "@mongodb-js/compass-sidebar": "^5.24.1",
-    "@mongodb-js/compass-welcome": "^0.22.1",
-    "@mongodb-js/compass-workspaces": "^0.5.1",
-    "@mongodb-js/connection-info": "^0.1.5",
-    "@mongodb-js/connection-storage": "^0.8.1",
-    "hadron-app-registry": "^9.1.8",
-    "hadron-ipc": "^3.2.12",
-    "mongodb-data-service": "^22.18.1"
-  },
-  "peerDependencies": {
-    "@mongodb-js/atlas-service": "^0.15.1",
-    "@mongodb-js/compass-aggregations": "^9.26.1",
-    "@mongodb-js/compass-app-stores": "^7.10.1",
-    "@mongodb-js/compass-collection": "^4.23.1",
-    "@mongodb-js/compass-components": "^1.22.1",
-    "@mongodb-js/compass-connections": "^1.25.1",
-    "@mongodb-js/compass-crud": "^13.24.1",
-    "@mongodb-js/compass-databases-collections": "^1.23.1",
-    "@mongodb-js/compass-explain-plan": "^6.24.1",
-    "@mongodb-js/compass-export-to-language": "^9.0.1",
-    "@mongodb-js/compass-field-store": "^9.0.19",
-    "@mongodb-js/compass-find-in-page": "^4.22.1",
-    "@mongodb-js/compass-import-export": "^7.23.1",
-    "@mongodb-js/compass-indexes": "^5.23.1",
-    "@mongodb-js/compass-logging": "^1.2.14",
-    "@mongodb-js/compass-query-bar": "^8.25.1",
-    "@mongodb-js/compass-saved-aggregations-queries": "^1.24.1",
-    "@mongodb-js/compass-schema": "^6.25.1",
-    "@mongodb-js/compass-schema-validation": "^6.24.1",
-    "@mongodb-js/compass-serverstats": "^16.23.1",
-    "@mongodb-js/compass-settings": "^0.26.1",
-    "@mongodb-js/compass-shell": "^3.23.1",
-    "@mongodb-js/compass-sidebar": "^5.24.1",
-    "@mongodb-js/compass-welcome": "^0.22.1",
-    "@mongodb-js/compass-workspaces": "^0.5.1",
-    "@mongodb-js/connection-info": "^0.1.5",
-    "@mongodb-js/connection-storage": "^0.8.1",
-    "hadron-app-registry": "^9.1.8",
-    "hadron-ipc": "^3.2.12",
-    "mongodb-data-service": "^22.18.1",
-    "react": "^17.0.2"
-  },
-  "devDependencies": {
-    "@electron/remote": "^2.1.2",
-    "@mongodb-js/eslint-config-compass": "^1.0.17",
-    "@mongodb-js/mocha-config-compass": "^1.3.7",
-    "@mongodb-js/my-queries-storage": "^0.5.1",
-    "@mongodb-js/prettier-config-compass": "^1.0.1",
-    "@mongodb-js/tsconfig-compass": "^1.0.3",
-    "@mongodb-js/webpack-config-compass": "^1.3.5",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/chai": "^4.2.21",
-    "chai": "^4.1.2",
-    "compass-preferences-model": "^2.18.1",
-    "depcheck": "^1.4.1",
-    "electron-mocha": "^12.2.0",
-    "eslint": "^7.25.0",
-    "eventemitter3": "^4.0.0",
-    "mocha": "^10.2.0",
-    "mongodb-ns": "^2.4.0",
-    "nyc": "^15.1.0",
-    "prettier": "^2.7.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "resolve": "^1.15.1",
-    "sinon": "^8.1.1",
-    "xvfb-maybe": "^0.2.1",
-    "mongodb": "^6.5.0",
-    "mongodb-build-info": "^1.7.0",
-    "resolve-mongodb-srv": "^1.1.2",
-    "mongodb-connection-string-url": "^2.6.0",
-    "mongodb-cloud-info": "^2.1.1"
-  },
-  "homepage": "https://github.com/mongodb-js/compass",
-  "bugs": {
-    "url": "https://jira.mongodb.org/projects/COMPASS/issues",
-    "email": "compass@mongodb.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mongodb-js/compass.git"
-  }
-}
