--- packages/compass-home/src/components/index.ts
@@ -1,3 +1,0 @@
-import Home from './home';
-
-export default Home;
