--- packages/compass-home/src/index.ts
@@ -1,20 +1,0 @@
-import { registerHadronPlugin } from 'hadron-app-registry';
-import Home from './components/home';
-
-export const CompassHomePlugin = registerHadronPlugin({
-  name: 'CompassHome',
-  component: Home,
-  activate(/* ..., { globalAppRegistry, localAppRegistry } */) {
-    // TODO: This is where we should be subscribing to appRegistry events
-    // instead of passing it directly to the Home component. Keeping it as-is
-    // as cleaning up compass-home is a bigger refactor that is not in scope
-    return {
-      store: {
-        state: {},
-      },
-      deactivate() {
-        /* noop */
-      },
-    };
-  },
-});
