--- packages/compass-home/tsconfig-lint.json
@@ -1,5 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["**/*"],
-  "exclude": ["node_modules", "dist"]
-}
