--- packages/compass-home/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@mongodb-js/tsconfig-compass/tsconfig.react.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["./src/**/*.spec.*"]
-}
