--- packages/compass-home/webpack.config.js
@@ -1,2 +1,0 @@
-const { compassPluginConfig } = require('@mongodb-js/webpack-config-compass');
-module.exports = compassPluginConfig;
