+++ packages/compass/package.json
@@ -179,14 +179,37 @@
     "@electron/rebuild": "^3.6.0",
     "@electron/remote": "^2.1.2",
     "@mongodb-js/atlas-service": "^0.15.1",
+    "@mongodb-js/compass-aggregations": "^9.26.1",
+    "@mongodb-js/compass-app-stores": "^7.10.1",
+    "@mongodb-js/compass-collection": "^4.23.1",
+    "@mongodb-js/compass-components": "^1.22.1",
+    "@mongodb-js/compass-connections": "^1.25.1",
+    "@mongodb-js/compass-crud": "^13.24.1",
+    "@mongodb-js/compass-databases-collections": "^1.23.1",
+    "@mongodb-js/compass-explain-plan": "^6.24.1",
+    "@mongodb-js/compass-export-to-language": "^9.0.1",
+    "@mongodb-js/compass-field-store": "^9.0.19",
+    "@mongodb-js/compass-find-in-page": "^4.22.1",
     "@mongodb-js/compass-generative-ai": "^0.8.1",
-    "@mongodb-js/compass-home": "^7.0.1",
+    "@mongodb-js/compass-import-export": "^7.23.1",
+    "@mongodb-js/compass-indexes": "^5.23.1",
+    "@mongodb-js/compass-intercom": "^0.2.1",
     "@mongodb-js/compass-logging": "^1.2.14",
+    "@mongodb-js/compass-query-bar": "^8.25.1",
+    "@mongodb-js/compass-saved-aggregations-queries": "^1.24.1",
+    "@mongodb-js/compass-schema": "^6.25.1",
+    "@mongodb-js/compass-schema-validation": "^6.24.1",
+    "@mongodb-js/compass-serverstats": "^16.23.1",
+    "@mongodb-js/compass-settings": "^0.26.1",
+    "@mongodb-js/compass-shell": "^3.23.1",
+    "@mongodb-js/compass-sidebar": "^5.24.1",
     "@mongodb-js/compass-utils": "^0.6.1",
+    "@mongodb-js/compass-welcome": "^0.22.1",
+    "@mongodb-js/compass-workspaces": "^0.5.1",
+    "@mongodb-js/connection-info": "^0.1.5",
     "@mongodb-js/connection-storage": "^0.8.1",
     "@mongodb-js/eslint-config-compass": "^1.0.17",
     "@mongodb-js/get-os-info": "^0.3.23",
-    "@mongodb-js/compass-intercom": "^0.2.1",
     "@mongodb-js/mocha-config-compass": "^1.3.7",
     "@mongodb-js/mongodb-downloader": "^0.2.1",
     "@mongodb-js/my-queries-storage": "^0.5.1",
@@ -218,13 +241,16 @@
     "marky": "^1.2.1",
     "mongodb": "^6.5.0",
     "mongodb-build-info": "^1.7.0",
+    "mongodb-cloud-info": "^2.1.1",
     "mongodb-connection-string-url": "^2.6.0",
     "mongodb-data-service": "^22.18.1",
     "mongodb-instance-model": "^12.18.1",
     "mongodb-log-writer": "^1.3.0",
+    "mongodb-ns": "^2.4.0",
     "node-fetch": "^2.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "resolve-mongodb-srv": "^1.1.3",
     "semver": "^7.5.4",
     "sinon": "^8.1.1",
     "source-code-pro": "^2.38.0",
