+++ packages/compass/src/app/components/workspace.tsx
