+++ packages/compass/src/app/utils/auto-connect.spec.ts
