+++ packages/compass/src/app/utils/csp.ts
