+++ packages/compass/src/app/utils/telemetry.ts
